// src/common/enums/social.enum.ts
export enum SocialMediaType {
  x = 'x',
  facebook = 'facebook',
  instagram = 'instagram',
  linkedin = 'linkedin',
  youtube = 'youtube',
  tiktok = 'tiktok',
  github = 'github',
  gitlab = 'gitlab',
  reddit = 'reddit',
  telegram = 'telegram',
  whatsapp = 'whatsapp',
  discord = 'discord',
}